Voice/PythonProject/requirements.txt
mediapipe==0.10.32
opencv-python==4.10.0.84
numpy==1.26.4
pyserial==3.5
SpeechRecognition==3.10.4
pyaudio==0.2.14; python_version < "3.12"
pygame==2.5.2
pyttsx3==2.90
pillow==10.4.0
